The "next event" string is now calculated
We don't always know exactly when the next event will be, but we usually know
something about it. More importantly, we rarely update the website once one
event is over, so the website is usually actually showing the date of the last
event instead of the next one.

To mitigate this, we now have an array of "future dates" for the next few
events, and the website is always going to show the "next event" from that
array that isn't a date in the past. If all the dates are in the past, instead
of showing a date, we'll say "a definir".

Closes: #1